4 modified

+++ Unity/Practice/Diablo/Assets/Scripts/InGame/Player/PlayerStatus.cs
@@ -15,7 +15,7 @@ namespace InGame.Player {
         [HideInInspector] public int dexterity;
         [HideInInspector] public int agility;
 
-        [HideInInspector] public int playerLV = 1;
+        [HideInInspector] public byte playerLV = 1;
         [HideInInspector] public float playerEXP;
         [HideInInspector] public int skillPoint;
